pyproject.toml
[tool.poetry]
name = "crowdreviews"
version = "0.1.0"
description = ""
authors = ["Chintal <chintaldraval@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-environ = "^0.9.0"
Django = "3.2.5"
graphene-django = "^2.15.0"
django-graphql-jwt = "0.3.0"
django-graphql-auth = "^0.3.16"
django-cors-headers = "3.7.0"
pip = "^22.2.2"
psycopg2-binary = "^2.9.3"
opentracing = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
